Use the compiler define _WIN32
This was causing weirdness because apparently MSYS2, etc does not
define WIN32 by default. So depending on which header was included
before libzippp, it might or might not be defined. Which was causing a
mismatch with the compiled library, which never had WIN32 defined. On
the other hand, _WIN32 is defined by the compiler, so that always
seems to do the right thing.